[OVN] Make the Logical_Switch_Port UP event more robust
In ML2/OVN it is possible that the OVN databases are configured in a
OVSDB RAFT cluster. If by any circumstance the RAFT leader changes, the
active IDL connections to OVN NB and SB databases will temporarily
disconnect and reconnect again. During this time gap, if the
``LogicalSwitchPortUpdateUpEvent`` arrives, it won't be processed.

This patch keeps the ``LogicalSwitchPortCreateEvent`` in the watch list.
Now if the IDL reconnects, all Logical_Switch_Port registers will be
received as CREATE events. If a ``LogicalSwitchPortUpdateUpEvent`` was
missed, now it will be processed.

Closes-Bug: #2101140

+++ neutron/tests/functional/base.py
@@ -396,7 +396,7 @@ class TestOVNFunctionalBase(test_plugin.Ml2PluginV2TestCase,
             except Exception:  # pylint:disable=bare-except
                 pass
 
-    def restart(self):
+    def restart(self, delete_dbs=True):
         self.stop()
 
         if self.ovsdb_server_mgr:
@@ -404,7 +404,8 @@ class TestOVNFunctionalBase(test_plugin.Ml2PluginV2TestCase,
         if self.ovn_northd_mgr:
             self.ovn_northd_mgr.stop()
 
-        self.ovsdb_server_mgr.delete_dbs()
+        if delete_dbs:
+            self.ovsdb_server_mgr.delete_dbs()
         self._start_ovsdb_server()
         self._start_idls()
         self._start_ovn_northd()
